% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.R
\name{Bayesian}
\alias{Bayesian}
\title{Comparison of Bayesian and classical MLR}
\usage{
Bayesian(model, data, alpha, burn = 1000, mc = 10000)
}
\arguments{
\item{model}{model}

\item{data}{dataset}

\item{alpha}{confidence level}

\item{burn}{burnin}

\item{mc}{mcmc}
}
\value{
Bayesian estimates, classical estimates, posterior plots and diagnostics
}
\description{
This function estimates the parameters in a multiple linear regression model using Bayesian and classical analysis.
}
\examples{
quasar<-readxl::read_excel("C:/Users/hstoh/OneDrive - University of Oklahoma/Fall 2020/MATH 5773/Data/QUASAR.xls")
Bayesian(model=RFEWIDTH~REDSHIFT+LINEFLUX+LUMINOSITY+AB1450,data=quasar,alpha=.95)
bub<-readxl::read_excel("C:/Users/hstoh/OneDrive - University of Oklahoma/Fall 2020/MATH 5773/Data/BUBBLE2.xls")
Bayesian(model=Density~Diameter+HeatFlux+MassFlux,data=bub,alpha=.90)

}
