% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{Bootstrap}
\alias{Bootstrap}
\title{Bootstrap MLR}
\usage{
Bootstrap(alpha, iter, model, data)
}
\arguments{
\item{alpha}{confidence level}

\item{iter}{iterations}

\item{model}{MLR model}

\item{data}{dataset}
}
\value{
Bootstrap beta estimates, CIs, and plots of point estimates; MLR summary and CIs
}
\description{
This function uses bootstrapping to create estimates and confidence intervals for the parameters of a multiple linear regression model.
}
\examples{
quasar<-readxl::read_excel("C:/Users/hstoh/OneDrive - University of Oklahoma/Fall 2020/MATH 5773/Data/QUASAR.xls")
Bootstrap(alpha=.95,iter=10000,model=RFEWIDTH~REDSHIFT+LINEFLUX+LUMINOSITY+AB1450,data=quasar)
bub<-readxl::read_excel("C:/Users/hstoh/OneDrive - University of Oklahoma/Fall 2020/MATH 5773/Data/BUBBLE2.xls")
Bootstrap(alpha=.9,iter=1000,model=Density~Diameter+HeatFlux+MassFlux,data=bub)

}
