% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validity.R
\name{validitycheck}
\alias{validitycheck}
\title{Check MLR assumptions}
\usage{
validitycheck(model1)
}
\arguments{
\item{model1}{MLR model}
}
\value{
mean of the residuals, residuals vs. fitted plot, normality check, outliers
}
\description{
This function can be used to check the assumptions of a multiple linear regression model.
}
\examples{
library(s20x)
data(camplake.df)
camp.lm=lm(Length~Age+Scale.Radius,data=camplake.df)
validitycheck(model1=camp.lm)
data(books.df)
book.lm=lm(length~book,data=books.df)
validitycheck(model1=book.lm)

}
