% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validity.R
\name{validitycheck}
\alias{validitycheck}
\title{Check MLR assumptions}
\usage{
validitycheck(model1)
}
\arguments{
\item{model1}{MLR model}
}
\value{
mean of the residuals, residuals vs. fitted plot, normality check, outliers
}
\description{
This function can be used to check the assumptions of a multiple linear regression model.
}
\examples{
quasar<-readxl::read_excel("C:/Users/hstoh/OneDrive - University of Oklahoma/Fall 2020/MATH 5773/Data/QUASAR.xls")
quasar.lm=lm(RFEWIDTH~REDSHIFT+LINEFLUX+LUMINOSITY+AB1450,data=quasar)
validitycheck(model1=quasar.lm)
bub<-readxl::read_excel("C:/Users/hstoh/OneDrive - University of Oklahoma/Fall 2020/MATH 5773/Data/BUBBLE2.xls")
bub.lm=lm(Density~Diameter+HeatFlux+MassFlux,data=bub)
validitycheck(model1=bub.lm)
}
